arrowutils: introduce package and move builder.GetValue to arrowutils
