+++ pqarrow/builder/optbuilders_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/apache/arrow/go/v8/arrow/memory"
 	"github.com/stretchr/testify/require"
 
+	"github.com/polarsignals/frostdb/pqarrow/arrowutils"
 	"github.com/polarsignals/frostdb/pqarrow/builder"
 )
 
@@ -51,7 +52,7 @@ func TestRepeatLastValue(t *testing.T) {
 		require.Equal(t, tc.b.Len(), 10)
 		a := tc.b.NewArray()
 		for i := 0; i < a.Len(); i++ {
-			v, err := builder.GetValue(a, i)
+			v, err := arrowutils.GetValue(a, i)
 			require.NoError(t, err)
 			require.Equal(t, tc.v, v)
 		}
