+++ pqarrow/builder/utils.go
@@ -126,26 +126,3 @@ func AppendGoValue(cb ColumnBuilder, v any) error {
 	}
 	return nil
 }
-
-// GetValue returns the value at index i in arr. If the value is null, nil is
-// returned.
-func GetValue(arr arrow.Array, i int) (any, error) {
-	if arr.IsNull(i) {
-		return nil, nil
-	}
-
-	switch a := arr.(type) {
-	case *array.Binary:
-		return a.Value(i), nil
-	case *array.FixedSizeBinary:
-		return a.Value(i), nil
-	case *array.String:
-		return a.Value(i), nil
-	case *array.Int64:
-		return a.Value(i), nil
-	case *array.Boolean:
-		return a.Value(i), nil
-	default:
-		return nil, fmt.Errorf("unsupported type for GetValue %T", a)
-	}
-}
